### 2020-01-27 #2042 @whedon
[REVIEW]: MetSim: A Python package for estimation anddisaggregation of meteorological data
**Submitting author:** @arbennett (<a href="http://orcid.org/0000-0002-7742-3138">Andrew Bennett</a>)
**Repository:** <a href="https://github.com/UW-Hydro/MetSim" target ="_blank">https://github.com/UW-Hydro/MetSim</a>
**Version:** v2.2.1
