### 2020-02-07 UW-Hydro/MetSim#200 @Chilipp
Time information doubled
This issue is part of the JOSS review in https://github.com/openjournals/joss-reviews/issues/2042

When running `ms examples/example_nc.conf`, the time information in the data is doubled (i.e. time-steps appear twice) and the output has more than 24 time steps per day.

Here is the code to reproduce the issue:

```bash
conda create -n metsim metsim cdo -c conda-forge && conda activate metsim
git clone https://github.com/UW-Hydro/MetSim.git && cd MetSim
ms examples/example_nc.conf --verbose
cdo tinfo forcing_19500101-19500131.nc | head -n 20
```

<summary>This is the output of the last CDO command</summary>

```

   Time axis :  1488 steps
     RefTime =  2000-01-01 00:00:00  Units = hours  Calendar = standard

         Verification Time
Timestep YYYY-MM-DD hh:mm:ss   Increment
     1   1950-01-01 00:00:00   --------
     2   1950-01-01 01:00:00   1 hour
     3   1950-01-01 01:00:00   0 seconds
     4   1950-01-01 02:00:00   1 hour
     5   1950-01-01 02:00:00   0 seconds
     6   1950-01-01 03:00:00   1 hour
     7   1950-01-01 03:00:00   0 seconds
     8   1950-01-01 04:00:00   1 hour
     9   1950-01-01 04:00:00   0 seconds
    10   1950-01-01 05:00:00   1 hour
    11   1950-01-01 05:00:00   0 seconds
    12   1950-01-01 06:00:00   1 hour
    13   1950-01-01 06:00:00   0 seconds
    14   1950-01-01 07:00:00   1 hour

```

Is this a feature or a bug?
