### 2020-02-07 UW-Hydro/MetSim#202 @Chilipp
Add API documentation for metsim.methods.mtclim
It seems like the `metsim.methods.mtclim` does not have any documentation although it defines quite a few important variables.

This issue is part of the JOSS review https://github.com/openjournals/joss-reviews/issues/2042
