### 2020-02-07 #2042 @Chilipp
Hey everyone! My apologies for the delay! I am done with the review and I highly appreciate your work here! MetSim seems to be a very valuable tool for downscaling meterology and I hope that you will provide a methodologic paper (i.e. where you describe the scientific details of your method) at some point as well!

The code-base and the documentation of this package is in a good condition and I could install and run the software on my local laptop (Linux). I did not try it on OS X and Ubuntu, and the authors do not state anything about OS-specific requirements (which is something that could be added to the installation instructions). I came across https://github.com/UW-Hydro/MetSim/issues/200 and https://github.com/UW-Hydro/MetSim/pull/201 during the review, and besides that, I have the following comments with reference to the checklist above:

---
- [ ] **Performance**: If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)

In the paper you state:

> building upon previous tools by improving performance

In what sense does MetSim improve the performance compared to the *previous tools*?

---

The documentation of the `metsim.methods.mtclim` module is missing (see https://github.com/UW-Hydro/MetSim/issues/202). The same is true for the `metsim.physics.solar_geom` function.

---

There are tests but it is not described in the documentation, how to run them. This should be mentioned somewhere (e.g. the README file, in the installation instructions or in the Contribution guidelines). I also strongly recommend to measure the coverage of the tests (e.g. using [pytest-cov](https://pytest-cov.readthedocs.io/en/latest/) and some free service, such as [codecov.io](https://codecov.io/)).

---
I could not find them. Please add a [CONTRIBUTING](https://help.github.com/articles/setting-guidelines-for-repository-contributors/) file or something equivalent.

---
I am not sure @sjpfenninger what is required here for the software paper as the discipline is quite diverse. The authors state in the paper

> We have based ``MetSim`` on methods from the Mountain Microclimate Simulator (``MTCLIM``) and the forcing preprocessor that was built into the Variable Infiltration Capacity (``VIC``) hydrological model version 4.
> ``MetSim`` provides a modern workflow, building upon previous tools by improving performance, adding new IO routines, allowing for exact restarts, and providing an extensible architecture which can incorporate new features.
> We have designed ``MetSim`` to fit into the broader scientific Python ecosystem, building on popular packages such as ``xarray``, ``dask``, ``pandas``, and ``numba``.

Is this enough for a review here? This paragraph does not describe the state of the field, but at least it highlights what is special about MetSim compared to other tools (modern workflow, improved performance, new IO routines, etc., although the statement about *improving perfomance* has to be revised, see *Performance* above).
