### 2020-02-07 bc288c4
Add docstrings to mtclim functions
