### 2020-02-07 UW-Hydro/MetSim#200 @arbennett
Ah, this is definitely a bug! The `example_nc.conf` file should generate output at 30 minute timesteps, but because the units are encoded as `hours since ....` the half hour gets rounded off.
