+++ CONTRIBUTING.md
@@ -1,0 +1,24 @@
+# Contributing to MetSim
+
+## Bug reports and feature requests
+
+If you experience any bugs or have use cases we have not considered feel free to open an [issue](https://github.com/UW-Hydro/MetSim/issues).
+When opening a new issue we request that you:
+
+ * Mention any specific library versions
+ * Provide code snippets as necessary
+ * Provide sample datasets as necessary
+
+We are open to taking feature requests, but are unlikely to act quickly on them.
+If you would like to implement a feature we would be happy to include it in future releases.
+To contribute a new feature we ask you to submit a pull request.
+
+## Submitting a pull request
+
+Whether you are updating documentation, fixing bugs, or implementing new features we are excited to work together!
+When submitting a pull request we ask that you:
+
+* Use pep8 style guides
+* Provide documentation of any new features
+* Only submit a single fix/feature per pull-request
+* Update dependencies as necessary
