### 2020-02-11 #2042 @Chilipp
Hey @arbennett! Thanks for your quick actions!

> Regarding tests: we do show how to run the tests when installing from source

Can you point me to the URL? I could not find it, neither in the installation instructions in the [docs](https://metsim.readthedocs.io/en/develop/#installation) nor in the [README](https://github.com/UW-Hydro/MetSim/blob/master/README.md#installation)

> I added a bit to the performance claim, but the main takeaway is that we have implemented scalable parallelism

Thanks for the clarification!

Otherwise, only the part about *State of the field* is unclear for me (see my comment above @sjpfenninger)
