### 2020-02-11 #2042 @arbennett
@Chilipp I updated the readme so that the line containing `pytest --verbose` is a little bit more clearly highlighted.
