### 2020-02-19 6cd6f82
Update paper.md
