+++ paper/paper.md
@@ -37,6 +37,12 @@ A summary of the available output variables and tunable parameters are included 
 ![Figure 1: An example of ``MetSim`` input and output. The daily values are shown on the left are used as input along with a number of parameters to produce the hourly output shown on the right. The sample data used to generate this figure is included in the ``MetSim`` repository.](figure1.png)
 
 We have based ``MetSim`` on methods from the Mountain Microclimate Simulator (``MTCLIM``) and the forcing preprocessor that was built into the Variable Infiltration Capacity (``VIC``) hydrological model version 4 [@Bohn:2013; @Thornton:1999; @Liang:1994].
+The development of ``MetSim`` was motivated by gaps in the approaches of both ``MTCLIM`` and the ``VIC`` forcing preprocessor.
+The ``MTCLIM`` software has not been updated since 2003 and the software does not implement the forcing disaggregation routines which are necessary for models to run at sub-daily timesteps.
+Recent developments on the ``VIC`` model have removed the forcing preprocessor due to algorithmic changes in the model driver.
+We noticed that processors for forcing estimation and disaggregation usually were implemented as ad hoc solutions based on similar workflows and algorithms, and wanted to formalize the process of generating forcing data by developing ``MetSim``.
+At the same time, this provides a basis for others to expand a set of commonly available routines.
+
 ``MetSim`` provides a modern workflow, building upon previous tools by improving performance by adding scalable parallelism, adding new IO routines, allowing for exact restarts, and providing an extensible architecture which can incorporate new features.
 We have designed ``MetSim`` to fit into the broader scientific Python ecosystem, building on popular packages such as ``xarray`` [@xarray], ``dask``[@dask], ``pandas``[@pandas], and ``numba`` [@numba].
 
