### 2020-02-28 9874037
Fix for sunrise/sunset north/south of the polar circles in set_min_max_hour routine, thus, allowing latitudes north/south of the polar circles.
