### 2020-02-28 0d1737c
Added missing_value and fill_value attributes to the netCDF variable attributes. Fix to transform input parameters expected as float to floats.
