### 2020-02-28 0472d0b
Fix to transform input parameters expected as float to floats.
