### 2020-03-05 #2042 @dsryberg
Hi all! Sorry for the delay/confusion on my part. I have now also completed the review, but I am unable to adjust the checklist since I cannot complete this step:

> Be sure to accept the invite at this URL: https://github.com/openjournals/joss-reviews/invitations

I am told that the invitation has expired. Is this easy to fix?
