### 2020-03-05 #2042 @arfon
> I am told that the invitation has expired. Is this easy to fix?

I've just re-invited you. The link at https://github.com/openjournals/joss-reviews/invitations should work again now.
