### 2020-03-06 #2042 @dsryberg
> > I am told that the invitation has expired. Is this easy to fix?
>
> I've just re-invited you. The link at https://github.com/openjournals/joss-reviews/invitations should work again now.

Unfortunately I am still told "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account." Although I have also made sure I am logging into the account by trying this on three different devices
