### 2020-03-06 #2042 @dsryberg
Overall: I agree that MetSim is a well structured, well documented, and sufficiently tested package, and I would definitely support it being published via JOSS. I nevertheless had a few hiccups during the review/testing process, and I also have trouble understanding who are the intended users of MetSim and what issue it is meant to resolve. I also suggest reworking some of the example scripts, as aspects of these failed for me. Finally, I think more detail should be given regarding MetSim's  "Disaggregation of daily simulation values to sub-daily timesteps". I'll go into more detail below:

### Functionality
> * [x]  **Installation:** Does installation proceed as outlined in the documentation?
- Installation was no issue for me on both a Linux and Windows PC. Although, as a minor issue, you may want to complete the same list of packages seen in the README as in the environment.yml file.

> * [x]  **Functionality:** Have the functional claims of the software been confirmed?
1. I successfully ran the built-in tests, and all 9 passed. This is great, but, executing these tests took 11 minutes, which is in my opinion long considering that they should be run often during future development processes. Especially the tests which run the examples "example_nc.conf" and "constant_vars_nc.conf" took the most time, so perhaps these could be simplified? In any case, 6 of the 9 tests are simply running the examples and they only ensure that an output file has been created. This will catch any raised errors, but it will miss the more subtle "sneaky" errors. Therefore I would suggest improving these by also checking some of the values which are produced to ensure that the code is working as intended.
    - Additionally, when I ran the tests I also evaluated the code coverage, which came out to a total of 75%, which is quite good! Nevertheless, the code within the file "physics.py" had the lowest specific score of 34%. Since the functions in this file are integral to the operation of MetSim, perhaps additional tests could be added to increase the coverage here.

2. I then tried to follow the tutorials found at [MetSim-tutorial](https://github.com/UW-Hydro/MetSim-tutorial)
    1. Example 1 worked as expected
    2. Example 2 appeared to work but...
        - Interactive plots were not functional (likely an issue unrelated to MetSim)
        - Tried changing timestep to 15 minutes
            + MetSim appeared to work as intended
            - There seems to be a bug in the way timestamps are saved. This leads to an error when plotting in the example. It is unclear how to resolve this issue
    3. Example 3 failed due to an issue with pandas. Maybe a result of conda installing the newest pandas version (1.0)

> * [x]  **Performance:** If there are any performance claims of the software, have they been confirmed? (If there are no claims, please check off this item.)
- From my perspective, MetSim does not seem to consume an undue amount of resources and parallelizes well. This is also discussed sufficiently in the paper.

### Documentation
> * [x]  **A statement of need:** Do the authors clearly state what problems the software is designed to solve and who the target audience is?
- (Good) Statements are given in the README which relate to what MetSim is capable of
- (Missing) The existing problem which the software solves is not specifically stated in the README
- (Missing) The intended audience is also not specified either
- (Concern) I am not a part of the climate modelling community (so I may be missing context which is otherwise obvious), however I do not understand the need to a climate data disagregator which operates on daily data when sub-daily data is already readily available from many sources (see MERRA2, ERA5, ect...). It is not clear to me who is meant to benefit from this software.

> * [x]  **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
- The examples are included and furthermore seem quite detailed but they are in my opinion cryptic. Instead of beginning the examples with "# This is an example of an input file for MetSim", I would suggest an introductory description of what the example is meant to accomplish.
- Also, please see the above comment regarding providing more examples/tests which cover the functions in "physics.py"

> * [x]  **Functionality documentation:** Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?
- The documentation is very well done. Although I've noticed differences between the documentation available on readthedocs versus what I can build from the cloned repository. For example, when I build the docs from the repository (using simply "make html") I do not see any items under the "API Reference" section. Although I'm not sure if this issue was caused by the doc files in MetSim, or else by Sphinx.

> * [x]  **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?
- Yes, but see the notes above concerning Functionality

### Software paper
> * [ ]  **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
- I agree with the comment from @Chilipp that the state of the field is not well addressed. I also agree with @sjpfenninger, that, if there are not other modules to compare, against then this can be pointed out instead.

### Other thoughts
1. It should be stated why using MetSim to disaggregate daily weather variables is preferable to simply downloading and using RCM or GCM products which are already hourly (for example, MERRA, ERA5, SARAH, ect...)
2. When listing the disaggregation capabilities of MetSim in the README, a brief note should be given regarding the quality or source of the algorithms used. For example, there is a sophisticated algorithm behind the solar irradiance disaggregation, but the wind speed disaggregation is just a forward fill for each time step in the day (at least, as far as I can tell by snooping through the code). An unaware user might not understand this disparity.
3. I also have not found any information regarding spatial resolution. Are the models implemented in MetSim intended to be applicable on all spatial scales?
