### 2020-03-06 #2042 @arbennett
@dsryberg thanks for the thorough review! I have updated the text of the paper as well as the documentation [here](https://metsim.readthedocs.io/en/master/data.html#sub-daily-output) to reflect the algorithms used to disaggregate. The more recent version of the article proof contains additional text about the state of the field, particularly with respect to MTCLIM and VIC.

Regarding your point number 1, I want to be clear that we aren't trying to say that MetSim is preferable to something like ERA5, but is complementary. MetSim can be used to generate sub-daily timeseries when something like ERA5 cannot be used (pre-1950, smaller spatial scales, etc). This actually brings up your point 3 as well. MetSim can be used at any spatial scale or mesh arrangement. I have also added a sentence to the paper to address this.

I have also updated the tutorial setup so that the Pandas error no longer occurs.
