### 2020-03-09 #2042 @sjpfenninger
@arbennet: Thanks for your quick work in addressing reviewer comments. Based on the reviews, I think your paper still needs a tiny bit of work.

Basically, the first paragraph in your paper should explain to a general scientifically literate audience whom your package is for and what problem(s) it solves. Right now, you delve into jargon almost immediately.

Note the requirements stated on https://joss.readthedocs.io/en/latest/submitting.html:
> A summary describing the high-level functionality and purpose of the software for a diverse, _non-specialist_ audience.
> A clear Statement of Need that illustrates the research purpose of the software.
