### 2020-03-10 81af2d0
Rework paper intro
