### 2020-03-11 #2042 @dsryberg
@arbennett  - The updated descriptions of the disaggregation methods looks a lot better now. Also the paper introduction is more readable as well. I am definitely satisfied with these two updates. However, have you taken into consideration the other topics I previously listed?

Namely:

1. Test run time
2. physics.py testing coverage
3. Making the examples more digestible
4. The potential bug with timesteps shorter than 1 hour
