### 2020-03-11 83895b7
Clean up fix for polar latitudes, change interpolation times
