### 2020-03-11 #2042 @arbennett
@sjpfenninger - I just release a new version, 2.2.1, to fix the timestamp bug as well as another bug found by a user when simulating polar regions. Is is possible to update the version for this paper?

@dsryberg - Thanks for the followup. I have released a new version to deal with the bug you describe in 4, along with an additional bug.

On the topic of tests, we don't have much bandwidth to improve the tests at this time. Specifically, for the time it takes to run the tests, I agree it is longer than necessary, but only a minor detriment in our view. Improving the `physics.py` test coverage would be great as well, but we test this "by proxy" by ensuring that the output is close to a benchmark from the VIC4.2 preprocessor. While our testing isn't perfect, it is still orders of magnitude better than the previous test-free situation of the ad-hoc arrangements described in the paper. We have an open issue to this end ([found here](https://github.com/UW-Hydro/MetSim/issues/175)), and I have commented there to improve the coverage.

For the examples, this was the reason we came up with the tutorial in a separate repo with a Binder setup for interactive use, as well as a filmed walkthrough (which I finally just updated the tutorial repo to link to). It provides background as well as some further exposition about how/why things are set up.
