### 2020-03-12 #2042 @dsryberg
@arbennett Thanks for the detailed response. In my opinion, the bug I came across was the most crucial issue to fix, so thanks for dealing with that. I've also just checked out the video walkthrough, and it is definitely very informative and well made. Maybe eventually this information should be put into a text form (so that one doesn't need to search through a 1.2 hr video to find the information they need) but, for the time being, I think it is sufficient. For the other issues, I think opening an issue for future improvements to the testing and examples is a good solution for now. After all, continued development is within the spirit of opensource software :)

With that being said, I am satisfied with the state of this software and I support it being published within JOSS.
