### 2020-03-18 #2042 @arbennett
@dsryberg and @Chilipp thanks to both of you for the really constructive reviews! I definitely feel like this whole process has made MetSim more accessible and uncovered a couple of bugs that are now fixed.

@sjpfenninger - let me know if there's anything else to do whenever you get around to looking at this again.

@arfon (and really the rest of you) - I just wanted to say how nice this review format is, I really appreciate the workflows and how clearly laid out everything is. This is my first submission to JOSS and the experience has been really pleasant!
