### 2020-03-18 0fd32b4
Removing duplicate figure labels.
