### 2020-03-18 #2042 @arfon
@arbennett - I found an issue with your figure labels (Pandoc numbers these automatically). https://github.com/UW-Hydro/MetSim/pull/214 fixes this.

I'm not sure if @sjpfenninger is online currently but I'd like to give him a little time to wrap make his final checks on this before handing it off to the EiCs for publishing.
