### 2020-03-18 #2042 @sjpfenninger
@arbennett Apologies for the delay and thanks for your work addressing the reviewer comments. It looks like this is ready for acceptance. Just one quick question - you mentioned updating the version, but I'm not sure what you mean, can you clarify?
