### 2020-03-18 #2042 @arbennett
@sjpfenninger Up on the header of this review it has: Version: v2.2.0

I'm wondering if it's possible/necessary to change this to v2.2.1 since we've fixed some bugs since submission. It seems it's not really referenced in the paper or my submission on the JOSS website so maybe this isn't needed.
