### 2020-03-19 9c78e20
add template to map_blocks call
