+++ ci/requirements-3.6.yml
@@ -14,6 +14,8 @@ dependencies:
   - dask
   - distributed
   - toolz
+  - pip
   - pip:
     - coveralls
     - pytest-cov
+    - git+https://github.com/dcherian/xarray.git@map-blocks-schema
