### 2020-03-26 #2042 @sjpfenninger
@arbennett Almost there, thanks for your patience. I have updated the version to v2.2.1. I've also noticed a couple small issues with the bibliography:

* Can you fix the two references with missing DOIs?
* Can you check and ideally fix the ["null" page numbers](https://github.com/UW-Hydro/MetSim/blob/fbae8184a5d886e1cb0b488896dd3fe22c3d61af/paper/paper.bib#L121) on Bohn 2019?

Then, please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive. For the Zenodo/figshare archive, please make sure that:

* The title of the archive is the same as the JOSS paper title
* That the authors of the archive are the same as the JOSS paper authors

Once that's done, we can move forward with accepting this!
