+++ paper/paper.bib
@@ -4,6 +4,7 @@
   booktitle = { Proceedings of the 14th Python in Science Conference },
   pages     = { 130 - 136 },
   year      = { 2015 },
+  doi       = { 10.25080/Majora-7b98e3ed-013 },
   editor    = { Kathryn Huff and James Bergstra }
 }
 
@@ -20,6 +21,7 @@
   title     = { Data Structures for Statistical Computing in Python },
   booktitle = { Proceedings of the 9th Python in Science Conference },
   pages     = { 51 - 56 },
+  doi       = {  10.25080/Majora-92bf1922-00a },
   year      = { 2010 },
   editor    = { St\'efan van der Walt and Jarrod Millman }
 }
@@ -112,17 +114,15 @@
   keywords = "Global land surface modeling "
 }
 
-@article{Bohn:2019,
+@article{bohn2019,
   author = {Bohn, Theodore J. and Whitney, Kristen M. and Mascaro, Giuseppe and Vivoni, Enrique R.},
   title = {A Deterministic Approach for Approximating the Diurnal Cycle of Precipitation for Use in Large-Scale Hydrological Modeling},
   journal = {Journal of Hydrometeorology},
-  volume = {0},
-  number = {0},
-  pages = {null},
+  volume = {20},
+  number = {2},
+  pages = {297-317},
   year = {2019},
   doi = {10.1175/JHM-D-18-0203.1},
-  URL = {https://doi.org/10.1175/JHM-D-18-0203.1},
-  eprint = {https://doi.org/10.1175/JHM-D-18-0203.1}
 }
 
 @article{Cheng:2020,
