### 2020-03-26 #2042 @arbennett
@sjpfenninger - thanks again for your efforts in editing this paper. I've updated the DOIs for the dask and pandas citations. I've also uploaded to Zenodo. The DOI is 10.5281/zenodo.3728015
