+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @inproceedings{dask,
   author    = { Matthew Rocklin },
   title     = { Dask: Parallel Computation with Blocked algorithms and Task Scheduling },
-  booktitle = { Proceedings of the 14th Python in Science Conference },
+  booktitle = {{Proceedings of the 14th Python in Science Conference}},
   pages     = { 130 - 136 },
   year      = { 2015 },
   doi       = { 10.25080/Majora-7b98e3ed-013 },
@@ -18,8 +18,8 @@
 
 @InProceedings{ pandas,
   author    = { Wes McKinney },
-  title     = { Data Structures for Statistical Computing in Python },
-  booktitle = { Proceedings of the 9th Python in Science Conference },
+  title     = { Data Structures for Statistical Computing in {P}ython },
+  booktitle = {{Proceedings of the 9th Python in Science Conference}},
   pages     = { 51 - 56 },
   doi       = {  10.25080/Majora-92bf1922-00a },
   year      = { 2010 },
@@ -28,14 +28,14 @@
 
 @inproceedings{numba,
  author = {Lam, Siu Kwan and Pitrou, Antoine and Seibert, Stanley},
- title = {Numba: A LLVM-Based Python JIT Compiler},
+ title = {Numba: A {LLVM}-Based {P}ython {JIT} Compiler},
  year = {2015},
  isbn = {9781450340052},
  publisher = {Association for Computing Machinery},
  address = {New York, NY, USA},
  url = {https://doi.org/10.1145/2833157.2833162},
  doi = {10.1145/2833157.2833162},
- booktitle = {Proceedings of the Second Workshop on the LLVM Compiler Infrastructure in HPC},
+ booktitle = {Proceedings of the Second Workshop on the {LLVM} Compiler Infrastructure in {HPC}},
  articleno = {Article 7},
  numpages = {6},
  keywords = {compiler, Python, LLVM},
@@ -96,7 +96,7 @@
 }
 
 @article{Bohn:2013,
-  title = "Global evaluation of MTCLIM and related algorithms for forcing of ecological and hydrological models ",
+  title = "Global evaluation of {MTCLIM} and related algorithms for forcing of ecological and hydrological models ",
   journal = "Agricultural and Forest Meteorology ",
   volume = "176",
   number = "",
@@ -126,8 +126,8 @@
 }
 
 @article{Cheng:2020,
-  title={Reservoirs modify river thermal regime sensitivity to climate change: a case study in the southeastern United States},
-  journal={Water Resources Research, In Review},
+  title={Reservoirs modify river thermal regime sensitivity to climate change: a case study in the southeastern {U}nited {S}tates. (In review)},
+  journal={Water Resources Research},
   author={Cheng, Yifan and Voisin, Nathalie and Yearsley, John and Nijssen, Bart},
   year={2020}
 }
