### 2020-03-27 #2042 @danielskatz
@arbennett - I've suggested some changes in https://github.com/UW-Hydro/MetSim/pull/216 and https://github.com/UW-Hydro/MetSim/pull/217
